+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 ## [Unreleased]
 
 
+## [v355] - 2026-03-27
+
+- Ruby 3.2.11 is now available
+
+
+## [v354] - 2026-03-26
+
+- Ruby 3.3.11 is now available
+
+
+## [v353] - 2026-03-25
+
+- Add heroku-26 stack support
+
 ## [v352] - 2026-03-16
 
 - Ruby 4.0.2 is now available
@@ -1923,7 +1937,10 @@ Bugfixes:
 * Change gem detection to use lockfile parser
 * use `$RACK_ENV` when thin is detected for rack apps
 
-[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v352...main
+[unreleased]: https://github.com/heroku/heroku-buildpack-ruby/compare/v355...main
+[v355]: https://github.com/heroku/heroku-buildpack-ruby/compare/v354...v355
+[v354]: https://github.com/heroku/heroku-buildpack-ruby/compare/v353...v354
+[v353]: https://github.com/heroku/heroku-buildpack-ruby/compare/v352...v353
 [v352]: https://github.com/heroku/heroku-buildpack-ruby/compare/v351...v352
 [v351]: https://github.com/heroku/heroku-buildpack-ruby/compare/v350...v351
 [v350]: https://github.com/heroku/heroku-buildpack-ruby/compare/v349...v350